+++ tmux/aliases.zsh
@@ -1,0 +1,6 @@
+# tmux aliases
+alias t='tmux new-session -A -s main'
+alias ta='tmux attach -t'
+alias tls='tmux list-sessions'
+alias tks='tmux kill-session -t'
+alias tka='tmux kill-server'